requirements.txt
#python-irodsclient>=1.0.0
Django>=4.2,<5.0 # LTS
django-filter>=24.0,<25.0
django-bootstrap4>=24.0,<25.0
django-q>=1.3.0,<2.0
django-tables2>=2.4.0,<3.0
django-view-breadcrumbs>=2.1.0,<3.0
psycopg2>=2.8.0,<3.0
gunicorn>20.0.0
pandas>=1.2.0,<2.0
numpy>=1.20.0,<2.0
openpyxl>=3.0.0,<4.0
xlsxwriter>=1.3.0,<2.0